Modularize code and update dependencies
- Rename ContentProvider to TwigContentProvider
- Extract ContentProvider interface
- Split TwigContentProvider into actual content provider and factory, so
   the unit tests don't have to use reflection.
- Define configuration value object for factory - less code and tests
  needed for checking array values.

Update to PHPUnit 10.

This is a BC break because ContentProvider is now an interface and the
construction has changed.

Remove release notes from README, we're using the GitHub releases page
instead.
