+++ .travis.yml
@@ -3,7 +3,7 @@ language: php
 matrix:
   fast_finish: true
   include:
-  - php: 8.0
+  - php: 8.1
 
 sudo: false
