+++ composer.json
@@ -1,16 +1,16 @@
 {
     "name": "wmde/fundraising-content-provider",
     "description": "WMDE fundraising application content provider and filter",
-    "license": "GPL-2.0+",
+    "license": "GPL-2.0-or-later",
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "ezyang/htmlpurifier": "^4.9",
-        "twig/twig": "~3.0",
+        "twig/twig": "~3.4",
         "symfony/console": "^5.4 || ^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~9.5",
-        "wmde/fundraising-phpcs": "~5.0",
+        "phpunit/phpunit": "~10.0",
+        "wmde/fundraising-phpcs": "~7.0",
         "mikey179/vfsstream": "^1.6"
     },
     "repositories": [
