+++ phpunit.xml.dist
@@ -1,19 +1,16 @@
 <?xml version="1.0"?>
-<phpunit
+<phpunit 
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd"
+	xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/10.0/phpunit.xsd"
 	backupGlobals="false"
-	backupStaticAttributes="false"
 	colors="true"
 	beStrictAboutTestsThatDoNotTestAnything="true"
-	convertErrorsToExceptions="true"
-	convertNoticesToExceptions="true"
-	convertWarningsToExceptions="true"
 	stopOnError="false"
 	stopOnFailure="false"
 	stopOnIncomplete="false"
 	stopOnSkipped="false"
-	verbose="false"
+	cacheDirectory=".phpunit.cache"
+	backupStaticProperties="false"
 	>
   <coverage includeUncoveredFiles="true">
     <include>
